+++ spec/dummy/app/models/multi_applications_account.rb
@@ -1,0 +1,3 @@
+class MultiApplicationsAccount < ActiveRecord::Base
+  include BookingSync::Engine::Models::MultiApplicationsAccountModel
+end