+++ spec/dummy/db/migrate/20181128184412_add_host_to_accounts.rb
@@ -1,5 +1,0 @@
-class AddHostToAccounts < ActiveRecord::Migration[5.1]
-  def change
-    add_column :accounts, :host, :string
-  end
-end