+++ spec/models/multi_application_account_spec.rb
@@ -1,6 +1,6 @@
 require "spec_helper"
 
-RSpec.describe Account, type: :model do
+RSpec.describe MultiApplicationsAccount, type: :model do
   shared_examples "it takes attributes from auth" do
     it "sets name" do
       expect(account.name).to eq "business name"
@@ -14,63 +14,73 @@ RSpec.describe Account, type: :model do
   end
 
   describe "validations" do
-    it { is_expected.to validate_uniqueness_of(:synced_id) }
+    let!(:account) do
+      MultiApplicationsAccount.create!(provider: "bookingsync", synced_id: 123, host: "example.test")
+    end
+
+    it { is_expected.to validate_uniqueness_of(:synced_id).scoped_to(:host).case_insensitive }
   end
 
   describe ".from_omniauth" do
     let(:auth) { OmniAuth.config.mock_auth[:bookingsync] }
+    let!(:account) do
+      MultiApplicationsAccount.create!(provider: "bookingsync", synced_id: 123, host: "example.test")
+    end
+    let!(:account2) do
+      MultiApplicationsAccount.create!(provider: "bookingsync", synced_id: 456, host: "example2.test")
+    end
 
-    before { Account.create!(provider: "bookingsync", synced_id: 456) }
-  
-    context "when account exists" do
-      let!(:account) { Account.create!(provider: "bookingsync", synced_id: 123) }
-  
-      context "with host given" do
-        it "loads the existing account" do
-          expect(Account.from_omniauth(auth, "example.test")).to eql(account)
-        end
-      end
-
-      context "without host given" do
-        it "loads the existing account" do
-          expect(Account.from_omniauth(auth, nil)).to eql(account)
-        end
-      end
-  
-      describe "the updated account" do
-        before do
-          Account.from_omniauth(auth, "example.test")
-          account.reload
-        end
-  
-        it_behaves_like "it takes attributes from auth"
+    context "without host given" do
+      it "raises an error" do
+        expect { MultiApplicationsAccount.from_omniauth(auth, nil) }.to raise_error(ArgumentError,
+          "The `host` variable must be passed when using BookingSync Engine with multi application support")
       end
     end
-  
-    context "when account doesn't exist" do
-      it "creates new account" do
-        expect {
-          Account.from_omniauth(auth, "example.test")
-        }.to change { Account.count }.by(1)
-      end
-  
-      describe "the newly created account" do
-        let!(:account) { Account.from_omniauth(auth, "example.test") }
-  
-        it "sets synced_id and provider from auth" do
-          expect(account.synced_id).to eq 123
-          expect(account.provider).to eq "bookingsync"
+
+    context "with host given" do
+      context "when account exists" do
+        it "loads the existing account" do
+          expect(MultiApplicationsAccount.from_omniauth(auth, "example.test")).to eql(account)
         end
-  
-        it_behaves_like "it takes attributes from auth"
+    
+        describe "the updated account" do
+          before do
+            MultiApplicationsAccount.from_omniauth(auth, "example.test")
+            account.reload
+          end
+    
+          it_behaves_like "it takes attributes from auth"
+        end
+      end
+    
+      context "when account doesn't exist" do
+        it "creates new account" do
+          expect {
+            MultiApplicationsAccount.from_omniauth(auth, "example3.test")
+          }.to change { MultiApplicationsAccount.count }.by(1)
+        end
+    
+        describe "the newly created account" do
+          let!(:account) { MultiApplicationsAccount.from_omniauth(auth, "example3.test") }
+    
+          it "sets synced_id and provider from auth as well as host" do
+            expect(account.synced_id).to eq 123
+            expect(account.provider).to eq "bookingsync"
+            expect(account.host).to eq "example3.test"
+          end
+    
+          it_behaves_like "it takes attributes from auth"
+        end
       end
     end
   end
 
   describe "#token" do
     let(:expires_at) { 1.day.from_now.to_i }
-    let!(:account) { Account.create!(synced_id: 123, oauth_access_token: "token",
-      oauth_refresh_token: "refresh_token", oauth_expires_at: expires_at) }
+    let!(:account) do
+      MultiApplicationsAccount.create!(synced_id: 123, oauth_access_token: "token",
+        oauth_refresh_token: "refresh_token", oauth_expires_at: expires_at, host: "example.test")
+    end
     let(:oauth_client) { double }
 
     before do
@@ -105,7 +115,7 @@ RSpec.describe Account, type: :model do
       end
   
       after do
-        Account.destroy_all
+        MultiApplicationsAccount.destroy_all
       end
   
       it "refreshes the token" do
@@ -118,40 +128,52 @@ RSpec.describe Account, type: :model do
   describe "#application_token" do
     let(:client_credentials) { double }
     let(:oauth2_client) { double }
-    let!(:account) { Account.create!(synced_id: 123) }
 
-    it "returns a client credential token setup without default params" do
+    let!(:account) { MultiApplicationsAccount.create!(synced_id: 123, host: "test.example") }
+    let!(:application) { Application.create!(host: "test.example", client_id: "123", client_secret: "456") }
+
+    it "returns a client credential token setup with application's credentials" do
       expect(client_credentials).to receive(:get_token).and_return("client_credentials_token")
       expect(oauth2_client).to receive(:client_credentials).and_return(client_credentials)
       expect(OAuth2::Client).to receive(:new)
-        .with(ENV['BOOKINGSYNC_APP_ID'], ENV['BOOKINGSYNC_APP_SECRET'], any_args)
+        .with(application.client_id, application.client_secret, any_args)
         .and_return(oauth2_client)
 
       expect(BookingSync::Engine).to receive(:application_token)
-        .with(no_args).at_least(1).and_call_original
+        .with(client_id: application.client_id, client_secret: application.client_secret)
+        .at_least(1)
+        .and_call_original
       expect(account.application_token).to eq "client_credentials_token"
     end
   end
 
   describe "#oauth_client" do
-    let!(:account) { Account.create!(synced_id: 123) }
+    let!(:account) { MultiApplicationsAccount.create!(synced_id: 123, host: "test.example") }
+    let!(:application) { Application.create!(host: "test.example", client_id: "123", client_secret: "456") }
 
-    it "returns a BookingSync::Engine.oauth_client setup without default params" do
-      expect(BookingSync::Engine).to receive(:oauth_client).with(no_args).and_call_original
+    it "returns a BookingSync::Engine.oauth_client setup with application's credentials" do
+      expect(BookingSync::Engine).to receive(:oauth_client)
+        .with(client_id: application.client_id, client_secret: application.client_secret)
+        .at_least(1)
+        .and_call_original
       expect(account.oauth_client).to be_an OAuth2::Client
+      expect(account.oauth_client.id).to eq application.client_id
+      expect(account.oauth_client.secret).to eq application.client_secret
     end
   end
 
   describe "#application" do
-    let!(:account) { Account.create!(synced_id: 123) }
+    let!(:account) { MultiApplicationsAccount.create!(synced_id: 123, host: "test.example") }
+    let!(:application) { Application.create!(host: "test.example", client_id: "123", client_secret: "456") }
+    let!(:application2) { Application.create!(host: "test2.example", client_id: "234", client_secret: "567") }
 
-    it "returns nil" do
-      expect(account.application).to be_nil
+    it "returns the application matching the host" do
+      expect(account.application).to eq application
     end
   end
 
   describe "#api" do
-    let!(:account) { Account.new }
+    let!(:account) { MultiApplicationsAccount.new }
     let(:oauth_client) { double }
 
     before do
@@ -170,8 +192,8 @@ RSpec.describe Account, type: :model do
 
   describe "#clear_token!" do
     it "clears token related fields on account" do
-      account = Account.create!(oauth_access_token: "token",
-        oauth_refresh_token: "refresh", oauth_expires_at: "expires")
+      account = MultiApplicationsAccount.create!(oauth_access_token: "token",
+        oauth_refresh_token: "refresh", oauth_expires_at: "expires", host: "example.test")
 
       account.clear_token!
       account.reload